refactor(app): extract generate gallery state
